test: reorganize pi-ref-tools tests by concern
