+++ packages/pi-ref-tools/__tests__/helpers.test.ts
@@ -1,30 +1,25 @@
-import { existsSync, mkdirSync, mkdtempSync, readFileSync, writeFileSync } from "node:fs";
-import { homedir, tmpdir } from "node:os";
-import { join, resolve } from "node:path";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { existsSync, readFileSync } from "node:fs";
+import { afterEach, describe, expect, it } from "vitest";
 import {
-  DEFAULT_CONFIG_FILE,
   formatToolOutput,
   isJsonRpcResponse,
   isRecord,
   normalizeNumber,
-  parseConfig,
-  parseTimeoutMs,
   redactApiKey,
-  resolveConfigPath,
   resolveEffectiveLimits,
   splitParams,
   toJsonString,
   writeTempFile,
 } from "../extensions/index.js";
 
-describe("pi-ref-tools helpers", () => {
+describe("pi-ref-tools helper utilities", () => {
   it("splits params and clamps limits", () => {
     const { mcpArgs, requestedLimits } = splitParams({
       piMaxBytes: "100",
       piMaxLines: 5,
       query: "hello",
     });
+
     expect(mcpArgs).toEqual({ query: "hello" });
     expect(requestedLimits).toEqual({ maxBytes: 100, maxLines: 5 });
 
@@ -32,31 +27,25 @@ describe("pi-ref-tools helpers", () => {
     expect(effective).toEqual({ maxBytes: 120, maxLines: 2 });
   });
 
-  it("resolves effective limits using defaults when not requested", () => {
-    const effective = resolveEffectiveLimits({}, { maxBytes: 51200, maxLines: 2000 });
-    expect(effective).toEqual({ maxBytes: 51200, maxLines: 2000 });
+  it("uses configured limits when no overrides are requested", () => {
+    expect(resolveEffectiveLimits({}, { maxBytes: 51200, maxLines: 2000 })).toEqual({
+      maxBytes: 51200,
+      maxLines: 2000,
+    });
   });
 
-  it("resolves effective limits with only bytes requested", () => {
-    const effective = resolveEffectiveLimits({ maxBytes: 100 }, { maxBytes: 51200, maxLines: 2000 });
-    expect(effective.maxBytes).toBe(100);
-    expect(effective.maxLines).toBe(2000);
+  it("supports byte-only or line-only overrides", () => {
+    expect(resolveEffectiveLimits({ maxBytes: 100 }, { maxBytes: 51200, maxLines: 2000 })).toEqual({
+      maxBytes: 100,
+      maxLines: 2000,
+    });
+    expect(resolveEffectiveLimits({ maxLines: 500 }, { maxBytes: 51200, maxLines: 2000 })).toEqual({
+      maxBytes: 51200,
+      maxLines: 500,
+    });
   });
 
-  it("resolves effective limits with only lines requested", () => {
-    const effective = resolveEffectiveLimits({ maxLines: 500 }, { maxBytes: 51200, maxLines: 2000 });
-    expect(effective.maxBytes).toBe(51200);
-    expect(effective.maxLines).toBe(500);
-  });
-
-  it("parses timeout values", () => {
-    expect(parseTimeoutMs("250", 10)).toBe(250);
-    expect(parseTimeoutMs("0", 10)).toBe(10);
-    expect(parseTimeoutMs(undefined, 30000)).toBe(30000);
-    expect(parseTimeoutMs("abc", 500)).toBe(500);
-  });
-
-  it("normalizes numbers from strings and numbers", () => {
+  it("normalizes numeric values", () => {
     expect(normalizeNumber(42)).toBe(42);
     expect(normalizeNumber("123")).toBe(123);
     expect(normalizeNumber("abc")).toBeUndefined();
@@ -70,343 +59,106 @@ describe("pi-ref-tools helpers", () => {
     expect(redactApiKey("short")).toBe("***");
     expect(redactApiKey("abcdefghijklmnop")).toBe("abcd...mnop");
   });
-
-  it("keeps DEFAULT_CONFIG_FILE available for reference", () => {
-    expect(DEFAULT_CONFIG_FILE).toMatchObject({
-      url: "https://api.ref.tools/mcp",
-      maxBytes: 51200,
-      maxLines: 2000,
-    });
-  });
-
-  it("loads config from standard pi settings files", async () => {
-    const originalHome = process.env.HOME;
-    const originalCwd = process.cwd();
-    const tempHome = mkdtempSync(join(tmpdir(), "pi-ref-settings-home-"));
-    const tempProject = mkdtempSync(join(tmpdir(), "pi-ref-settings-project-"));
-
-    mkdirSync(join(tempHome, ".pi", "agent"), { recursive: true });
-    mkdirSync(join(tempProject, ".pi"), { recursive: true });
-
-    writeFileSync(
-      join(tempHome, ".pi", "agent", "settings.json"),
-      JSON.stringify({ "pi-ref-tools": { timeoutMs: 1234 } }),
-      "utf-8",
-    );
-    writeFileSync(
-      join(tempProject, ".pi", "settings.json"),
-      JSON.stringify({ "pi-ref-tools": { url: "https://project.example/mcp", maxLines: 99 } }),
-      "utf-8",
-    );
-
-    process.env.HOME = tempHome;
-    process.chdir(tempProject);
-    vi.resetModules();
-
-    try {
-      const mod = await import("../extensions/index.js");
-      const config = mod.loadRuntimeConfig({ getFlag: () => undefined } as never);
-      expect(config?.apiKey).toBeUndefined();
-      expect(config?.timeoutMs).toBe(1234);
-      expect(config?.url).toBe("https://project.example/mcp");
-      expect(config?.maxLines).toBe(99);
-    } finally {
-      process.chdir(originalCwd);
-      if (originalHome) process.env.HOME = originalHome;
-      else delete process.env.HOME;
-    }
-  });
-
-  it("warns when a legacy config file exists but is ignored", async () => {
-    const originalHome = process.env.HOME;
-    const originalCwd = process.cwd();
-    const tempHome = mkdtempSync(join(tmpdir(), "pi-ref-legacy-home-"));
-    const tempProject = mkdtempSync(join(tmpdir(), "pi-ref-legacy-project-"));
-
-    mkdirSync(join(tempHome, ".pi", "agent", "extensions"), { recursive: true });
-    writeFileSync(
-      join(tempHome, ".pi", "agent", "extensions", "ref-tools.json"),
-      JSON.stringify({ apiKey: "legacy-key" }),
-      "utf-8",
-    );
-
-    process.env.HOME = tempHome;
-    process.chdir(tempProject);
-    vi.resetModules();
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => undefined);
-
-    try {
-      const mod = await import("../extensions/index.js");
-      expect(mod.loadRuntimeConfig({ getFlag: () => undefined } as never)).toBeNull();
-      expect(warnSpy).toHaveBeenCalledWith(expect.stringContaining("Ignoring legacy config file"));
-    } finally {
-      warnSpy.mockRestore();
-      process.chdir(originalCwd);
-      if (originalHome) process.env.HOME = originalHome;
-      else delete process.env.HOME;
-    }
-  });
 });
 
 describe("pi-ref-tools type guards", () => {
-  it("isRecord returns true for plain objects", () => {
+  it("recognizes plain objects", () => {
     expect(isRecord({})).toBe(true);
     expect(isRecord({ a: 1 })).toBe(true);
     expect(isRecord({ nested: { deep: true } })).toBe(true);
   });
 
-  it("isRecord returns false for non-objects", () => {
+  it("rejects non-object values", () => {
     expect(isRecord(null)).toBe(false);
     expect(isRecord(undefined)).toBe(false);
     expect(isRecord("string")).toBe(false);
     expect(isRecord(123)).toBe(false);
     expect(isRecord(true)).toBe(false);
     expect(isRecord([])).toBe(false);
-    expect(isRecord([1, 2, 3])).toBe(false);
   });
 
-  it("isRecord returns true for objects with special keys", () => {
-    expect(isRecord({ __proto__: { admin: true } })).toBe(true);
-    expect(isRecord({ constructor: () => {} })).toBe(true);
-  });
-
-  it("isJsonRpcResponse returns true for valid responses", () => {
+  it("recognizes JSON-RPC responses", () => {
     expect(isJsonRpcResponse({ jsonrpc: "2.0", result: {} })).toBe(true);
     expect(isJsonRpcResponse({ jsonrpc: "2.0", id: 1, result: "test" })).toBe(true);
     expect(isJsonRpcResponse({ jsonrpc: "2.0", id: "abc", error: { code: -32600, message: "Invalid" } })).toBe(true);
-  });
-
-  it("isJsonRpcResponse returns false for invalid responses", () => {
     expect(isJsonRpcResponse({})).toBe(false);
     expect(isJsonRpcResponse({ jsonrpc: "1.0" })).toBe(false);
-    expect(isJsonRpcResponse({ jsonrpc: "2.0", id: 1 })).toBe(true); // passes isRecord and has jsonrpc 2.0
     expect(isJsonRpcResponse(null)).toBe(false);
-    expect(isJsonRpcResponse("string")).toBe(false);
-  });
-
-  it("isJsonRpcResponse handles edge cases", () => {
-    expect(isJsonRpcResponse({ jsonrpc: "2.0", error: { code: -32600 } })).toBe(true);
-    expect(isJsonRpcResponse({ jsonrpc: "2.0", result: null })).toBe(true);
   });
 });
 
-describe("pi-ref-tools toJsonString", () => {
-  it("returns strings as-is", () => {
+describe("pi-ref-tools serialization and formatting", () => {
+  it("returns strings as-is and stringifies other values", () => {
     expect(toJsonString("hello")).toBe("hello");
-    expect(toJsonString("")).toBe("");
-  });
-
-  it("stringifies objects", () => {
-    const result = toJsonString({ a: 1, b: 2 });
-    expect(JSON.parse(result)).toEqual({ a: 1, b: 2 });
-  });
-
-  it("converts primitives", () => {
     expect(toJsonString(42)).toBe("42");
     expect(toJsonString(true)).toBe("true");
     expect(toJsonString(null)).toBe("null");
-  });
-});
-
-describe("pi-ref-tools resolveConfigPath", () => {
-  it("resolves paths starting with ~/", () => {
-    const result = resolveConfigPath("~/.pi/config.json");
-    expect(result).toContain(homedir());
-    expect(result).toContain(".pi/config.json");
+    expect(JSON.parse(toJsonString({ a: 1, b: 2 }))).toEqual({ a: 1, b: 2 });
   });
 
-  it("resolves paths starting with ~", () => {
-    const result = resolveConfigPath("~/.pi/config.json");
-    expect(result).toContain(".pi/config.json");
-  });
-
-  it("returns absolute paths as-is", () => {
-    const absolute = "/absolute/path/to/config.json";
-    expect(resolveConfigPath(absolute)).toBe(absolute);
-  });
-
-  it("resolves relative paths from cwd", () => {
-    const result = resolveConfigPath("relative/path.json");
-    expect(result).toBe(resolve(process.cwd(), "relative/path.json"));
-  });
-
-  it("trims whitespace", () => {
-    const result = resolveConfigPath("  ~/.pi/config.json  ");
-    expect(result).toContain(".pi/config.json");
-  });
-});
-
-describe("pi-ref-tools parseConfig", () => {
-  it("parses valid config", () => {
-    const raw = {
-      url: "https://api.example.com/mcp",
-      apiKey: "secret-key",
-      timeoutMs: 15000,
-      protocolVersion: "2025-01-01",
-      maxBytes: 1024,
-      maxLines: 500,
-    };
-    const result = parseConfig(raw, "/path/to/config.json");
-    expect(result).toEqual({
-      url: "https://api.example.com/mcp",
-      apiKey: "secret-key",
-      timeoutMs: 15000,
-      protocolVersion: "2025-01-01",
-      maxBytes: 1024,
-      maxLines: 500,
-    });
-  });
-
-  it("normalizes string values", () => {
-    const raw = { url: "  https://api.example.com  ", apiKey: "  " };
-    const result = parseConfig(raw, "/path");
-    expect(result.url).toBe("https://api.example.com");
-    expect(result.apiKey).toBeUndefined(); // empty after trim
-  });
-
-  it("ignores null/undefined values", () => {
-    const raw = { url: null, apiKey: undefined, timeoutMs: NaN };
-    const result = parseConfig(raw, "/path");
-    expect(result.url).toBeUndefined();
-    expect(result.apiKey).toBeUndefined();
-    expect(result.timeoutMs).toBeUndefined(); // NaN is not finite
-  });
-
-  it("throws for non-object config", () => {
-    expect(() => parseConfig(null, "/path")).toThrow("Invalid Ref MCP config");
-    expect(() => parseConfig("string", "/path")).toThrow("Invalid Ref MCP config");
-    expect(() => parseConfig(123, "/path")).toThrow("Invalid Ref MCP config");
-  });
-
-  it("handles negative timeout", () => {
-    const raw = { timeoutMs: -100 };
-    const result = parseConfig(raw, "/path");
-    expect(result.timeoutMs).toBe(-100); // negative numbers are finite
-  });
-
-  it("handles zero timeout", () => {
-    const raw = { timeoutMs: 0 };
-    const result = parseConfig(raw, "/path");
-    expect(result.timeoutMs).toBe(0); // 0 is finite
-  });
-
-  it("normalizes whitespace in protocol version", () => {
-    const raw = { protocolVersion: "  2025-01-01  " };
-    const result = parseConfig(raw, "/path");
-    expect(result.protocolVersion).toBe("2025-01-01");
-  });
-});
-
-describe("pi-ref-tools formatToolOutput", () => {
-  it("formats simple text result", () => {
+  it("formats text output without truncation", () => {
     const result = formatToolOutput(
       "test_tool",
       "https://api.example.com",
       { content: [{ type: "text", text: "Hello World" }] },
       { maxBytes: 50000, maxLines: 2000 },
     );
+
     expect(result.text).toBe("Hello World");
     expect(result.details.truncated).toBe(false);
   });
 
-  it("formats result without content array", () => {
-    const result = formatToolOutput("test_tool", "https://api.example.com", { result: "data" } as unknown as Parameters<
-      typeof formatToolOutput
-    >[2]);
-    expect(result.text).toContain("data");
-  });
-
-  it("handles multiple content blocks", () => {
-    const result = formatToolOutput("test_tool", "https://api.example.com", {
-      content: [
-        { type: "text", text: "Block 1" },
-        { type: "text", text: "Block 2" },
-      ],
-    });
-    expect(result.text).toContain("Block 1");
-    expect(result.text).toContain("Block 2");
-  });
-
-  it("handles non-text content blocks", () => {
-    const result = formatToolOutput("test_tool", "https://api.example.com", {
+  it("formats fallback and mixed content blocks", () => {
+    const withoutContent = formatToolOutput("test_tool", "https://api.example.com", { result: "data" } as never);
+    const mixed = formatToolOutput("test_tool", "https://api.example.com", {
       content: [
         { type: "image", url: "https://example.com/img.png" },
         { type: "text", text: "Description" },
       ],
     });
-    expect(result.text).toContain("Description");
+
+    expect(withoutContent.text).toContain("data");
+    expect(mixed.text).toContain("Description");
   });
 
-  it("truncates when exceeding maxLines", () => {
-    const lines = Array.from({ length: 100 }, (_, i) => `Line ${i + 1}`).join("\n");
+  it("truncates long output and includes temp file details", () => {
+    const longText = Array.from({ length: 100 }, (_, index) => `Line ${index + 1}`).join("\n");
     const result = formatToolOutput(
       "test_tool",
       "https://api.example.com",
-      { content: [{ type: "text", text: lines }] },
+      { content: [{ type: "text", text: longText }] },
       { maxBytes: 100000, maxLines: 10 },
     );
+
     expect(result.details.truncated).toBe(true);
     expect(result.details.truncation?.truncatedBy).toBe("lines");
-  });
-
-  it("truncates when exceeding maxBytes", () => {
-    const longText = "a".repeat(10000);
-    const result = formatToolOutput(
-      "test_tool",
-      "https://api.example.com",
-      { content: [{ type: "text", text: longText }] },
-      { maxBytes: 1000, maxLines: 5000 },
-    );
-    expect(result.details.truncated).toBe(true);
-    expect(result.details.truncation?.truncatedBy).toBe("bytes");
-  });
-
-  it("includes temp file path when truncated", () => {
-    const longText = "x".repeat(10000);
-    const result = formatToolOutput(
-      "test_tool",
-      "https://api.example.com",
-      { content: [{ type: "text", text: longText }] },
-      { maxBytes: 100, maxLines: 10 },
-    );
     expect(result.details.tempFile).toBeDefined();
     expect(result.text).toContain("Full output saved to:");
   });
 });
 
-describe("pi-ref-tools writeTempFile", () => {
+describe("pi-ref-tools temp files", () => {
   const tempFiles: string[] = [];
 
-  afterEach(() => {
-    import("node:fs").then(({ unlinkSync }) => {
-      tempFiles.forEach((f) => {
-        try {
-          unlinkSync(f);
-        } catch {
-          // ignore cleanup errors
-        }
-      });
-    });
+  afterEach(async () => {
+    const { unlinkSync } = await import("node:fs");
+    for (const file of tempFiles.splice(0)) {
+      try {
+        unlinkSync(file);
+      } catch {
+        // ignore cleanup errors
+      }
+    }
   });
 
-  it("writes temp file and returns path", () => {
-    const path = writeTempFile("test_tool", "content here");
-    tempFiles.push(path);
-    expect(path).toContain("pi-ref-tools-test_tool");
-    expect(path).toContain(".txt");
-    expect(existsSync(path)).toBe(true);
-  });
-
-  it("sanitizes tool name", () => {
-    const path = writeTempFile("my-tool-123!@#", "content");
-    tempFiles.push(path);
-    expect(path).toContain("my-tool-123__");
-  });
-
-  it("writes correct content", () => {
+  it("writes temp files with sanitized names and content", () => {
     const content = "test content\nwith multiple lines";
-    const path = writeTempFile("test", content);
-    tempFiles.push(path);
-    expect(readFileSync(path, "utf-8")).toBe(content);
+    const filePath = writeTempFile("my-tool-123!@#", content);
+    tempFiles.push(filePath);
+
+    expect(filePath).toContain("pi-ref-tools-my-tool-123__");
+    expect(filePath).toContain(".txt");
+    expect(existsSync(filePath)).toBe(true);
+    expect(readFileSync(filePath, "utf-8")).toBe(content);
   });
 });
